+++ Changes.md
@@ -14,3 +14,8 @@
 ## 20220220:
 - Added option to set MTR ressource account credentials
 - Added option to update to latest online available MTR app version
+
+## 20230121:
+- Updated available MTR Themes for version 4.15.58.0
+- Added option to set MTR Teams App language
+- Added minor visual improvements to the menu